Make parallel make output work
Previously, when running make in parallel, the second part of the
output (the object size) was calculated when the compile finished. When
running make in parallel, this caused the output of multiple first
parts while the second part of the corresponding output wasn't printed
yet causing the output to become misaligned.